requirements.txt
--index-url https://download.pytorch.org/whl/cu128
--extra-index-url https://pypi.org/simple

anthropic==0.72.0
autoawq==0.2.9
bitsandbytes==0.46.0
bleak==0.22.3
chromadb==0.5.13
cryptography==46.0.3
discord==2.3.1
ffmpeg-python==0.2.0
Flask==3.0.0
Flask-Cors==5.0.0
Flask-Session==0.8.0
google-api-python-client==1.6.2
httplib2==0.22.0
http-constants==0.5.0
ld2410-ble==0.2.0
nltk==3.9.1
numpy==2.0.2
oauth2client==4.1.3
openai==2.6.1
openai-whisper==20250625
playsound==1.2.2
PyAudio==0.2.14
pydub==0.25.1
pypdf==5.1.0
pysbd==0.3.4
pytest==8.2.2
PyYAML==6.0.2
qwen_vl_utils==0.0.11
sounddevice==0.4.6
timm==1.0.16
torch==2.7.1
torchvision==0.22.1
trafilatura==1.12.1
transformers==4.56.2
whisper-mic==1.1.1
yt-dlp==2024.11.18
httpx==0.27.0
pgsql-mcp-server==1.4.1
